Add files via upload
Data that was analyzed. Found from Kaggle. 

About this file
Remote source:https://www.kaggle.com/alexteboul/diabetes-health-indicators-dataset-notebook

diabetes _ binary _ 5050split _ health _ indicators _ BRFSS2015.csv is a clean dataset of 70,692 survey responses to the CDC's BRFSS2015. It has an equal 50-50 split of respondents with no diabetes and with either prediabetes or diabetes. The target variable Diabetes_binary has 2 classes. 0 is for no diabetes, and 1 is for prediabetes or diabetes. This dataset has 21 feature variables and is balanced.